common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
authors = [{ name = "Donald Thompson", email = "witt3rd@witt3rd.com" }]
dependencies = [
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "loguru>=0.7.3",
    "mcp[cli]>=1.10.1",
    "jinja2>=3.1.6",
    "chromadb>=1.0.15",
    "sentence-transformers>=5.0.0",
]

# Force re-evaluation of environment
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]

# Optional: If common itself needs to reference other workspace members (unlikely for a 'common' lib)
# [tool.uv.sources]
# other_member = { workspace = true }
